upath: provide single UPath implementation
